Bump up to 0.3D Update core dependency to 0.10D

+++ src/main/java/me/theseems/tomshelby/economypack/api/transaction/types/DepositTransaction.java
@@ -15,5 +15,5 @@ public interface DepositTransaction extends Transaction {
    * Get user to deposit to
    * @return to user's id
    */
-  Integer getTo();
+  Long getTo();
 }
